ai_certificate/requirements.txt
albumentations==1.3.1

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
scikit-image==0.22.0
pdf2image==1.16.3

# OCR & Text
pytesseract==0.3.10
arabic-reshaper==3.0.0
python-bidi==0.4.2
pypdf2==3.0.1

# ML/AI
torch==2.1.0
transformers==4.35.2
datasets==2.15.0
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.25.1
scikit-learn==1.3.2
joblib==1.3.2

# Data Augmentation
imgaug==0.4.0
augraphy==8.2.6


# Utilities
redis==5.0.1
aiohttp==3.9.1
aiofiles==23.2.1
psutil==5.9.6
requests==2.31.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
torchvision==0.16.0
